refactor(cli): split god-package into focused files
Extract from cli.go into separate concerns:
- terminal.go: TTY detection, hidden/visible input, warning helpers
- vaultpath.go: vault path resolution, profile expansion, dir helpers
- unlock.go: vault unlocking, passphrase resolution, session TTL

cli.go reduced from 530 to 152 lines (71% reduction).

Closes #144
